More LemmyV1 changes
- Fix a bug where bottom bar sometimes appears after an image is opened and closed when it should be hidden
- Fix a bug where system gestures conflict with app gestures in some places

+++ api/src/main/kotlin/com/idunnololz/summit/api/dto/lemmy/v4/models/Comment.kt
@@ -117,7 +117,7 @@ data class Comment (
 
     /* The comment id. */
     @SerializedName("id")
-    val id: kotlin.Double,
+    val id: kotlin.Int,
 
     @SerializedName("updated_at")
     val updatedAt: kotlin.String? = null
